pyproject.toml
[project]
name = "23-5-team6-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.17.2",
    "fastapi>=0.128.0",
    "httpx>=0.27.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
    "pytest>=8.3.4",
    "email-validator>=2.2.0",
]
